refactor: nextstep.mvc.tobe 패키지 구조 변경
