+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -2,6 +2,8 @@ package nextstep.mvc;
 
 import nextstep.mvc.asis.Controller;
 import nextstep.mvc.tobe.*;
+import nextstep.mvc.tobe.exception.NotFoundHandlerException;
+import nextstep.mvc.tobe.view.View;
 import nextstep.utils.LoggingUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
