+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/exception/NotFoundHandlerException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.exception;
 
 public class NotFoundHandlerException extends RuntimeException {
     public static final String MESSAGE = "핸들러를 찾을 수 없습니다.";
