+++ slipp/src/main/java/slipp/controller/HomeController.java
@@ -1,7 +1,7 @@
 package slipp.controller;
 
 import nextstep.mvc.asis.Controller;
-import nextstep.mvc.tobe.JspView;
+import nextstep.mvc.tobe.view.JspView;
 import nextstep.mvc.tobe.ModelAndView;
 import nextstep.web.annotation.RequestMapping;
 import slipp.support.db.DataBase;
